backend/requirements.txt
aniso8601==9.0.1
Click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
cryptography==41.0.7
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
gunicorn==21.2.0
